refactor(pipelines): Extract pipeline modules from monolithic file
